+++ src/eca/features/chat.clj
@@ -1479,4 +1479,5 @@
         (send-chat-contents! messages chat-ctx)
         (lifecycle/send-content! chat-ctx :system (assoc-some {:type :metadata} :title title))
         (config/notify-selected-model-changed! (:model chat) db* messenger config)
+        (config/notify-selected-trust-changed! (:trust chat) db* messenger)
         {:found? true :chat-id chat-id :title title}))))